+++ src/f2_est.h
@@ -21,7 +21,7 @@
 
 /*
   File: f2_est.h
-  
+
   Description: Header file to a set of functions to compute the
   recombination fraction in F2 experimental populations. These
   functions contain the EM algorithms for all possible combination of
@@ -38,24 +38,24 @@
 #include <Rcpp.h>
 using namespace Rcpp;
 
-Rcpp::NumericVector est_rf_C_C(std::vector<int> k_sub,
-			       std::vector<int> k1_sub,
-			       int n_ind);
-Rcpp::NumericVector est_rf_C_D_43(std::vector<int> k_sub,
-				  std::vector<int> k1_sub,
-				  int n_ind);
-Rcpp::NumericVector est_rf_C_D_51(std::vector<int> k_sub,
-				  std::vector<int> k1_sub,
-				  int n_ind);
-Rcpp::NumericVector est_rf_D_D_43(std::vector<int> k_sub,
-				  std::vector<int> k1_sub,
-				  int n_ind);
-Rcpp::NumericVector est_rf_D_D_51(std::vector<int> k_sub,
-				  std::vector<int> k1_sub,
-				  int n_ind);
-Rcpp::NumericVector est_rf_D_D_43_51(std::vector<int> k_sub,
-				     std::vector<int> k1_sub,
-				     int n_ind);
-Rcpp::NumericVector est_rf_A_A(std::vector<int> k_sub,
-			       std::vector<int> k1_sub,
-			       int n_ind);
+Rcpp::NumericVector est_rf_C_C(std::vector<long> k_sub,
+			       std::vector<long> k1_sub,
+			       long n_ind);
+Rcpp::NumericVector est_rf_C_D_43(std::vector<long> k_sub,
+				  std::vector<long> k1_sub,
+				  long n_ind);
+Rcpp::NumericVector est_rf_C_D_51(std::vector<long> k_sub,
+				  std::vector<long> k1_sub,
+				  long n_ind);
+Rcpp::NumericVector est_rf_D_D_43(std::vector<long> k_sub,
+				  std::vector<long> k1_sub,
+				  long n_ind);
+Rcpp::NumericVector est_rf_D_D_51(std::vector<long> k_sub,
+				  std::vector<long> k1_sub,
+				  long n_ind);
+Rcpp::NumericVector est_rf_D_D_43_51(std::vector<long> k_sub,
+				     std::vector<long> k1_sub,
+				     long n_ind);
+Rcpp::NumericVector est_rf_A_A(std::vector<long> k_sub,
+			       std::vector<long> k1_sub,
+			       long n_ind);
